# internal/delivery/http/handler/auth_dto.go
package handler

import "github.com/golang-jwt/jwt/v5"

type registerRequest struct {
	Username string `json:"username" binding:"required"`
	Password string `json:"password" binding:"required,min=6"`
	Name     string `json:"name" binding:"required"`
}

type registerResponse struct {
	ID       string `json:"id"`
	Username string `json:"username"`
	Name     string `json:"name"`
	Role     string `json:"role"`
	Message  string `json:"message"`
}

type loginRequest struct {
	Username string `json:"username" binding:"required"`
	Password string `json:"password" binding:"required,min=6"`
}

type loginResponse struct {
	AccessToken  string `json:"accessToken"`
	RefreshToken string `json:"refreshToken"`
}

type refreshResponse struct {
	AccessToken  string `json:"access_token" example:"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."`
	RefreshToken string `json:"refresh_token" example:"d3f78a9c2b4e8f110a1290bc5e6f7a8b9c0d1e2f3a4b5c6d7e8f9a0b1c2d3e4f"`
}

type AccessClaimsRequest struct {
	Username string `json:"username"`
	Role     string `json:"role"`
	jwt.RegisteredClaims
}

type RefreshClaimsRequest struct {
	Username string `json:"username"`
	jwt.RegisteredClaims
}